+++ src/Products/PloneMeeting/MeetingItem.py
@@ -75,6 +75,7 @@ from Products.PloneMeeting.browser.itemvotes import next_vote_is_linked
 from Products.PloneMeeting.config import AddAdvice
 from Products.PloneMeeting.config import AUTO_COPY_GROUP_PREFIX
 from Products.PloneMeeting.config import BUDGETIMPACTEDITORS_GROUP_SUFFIX
+from Products.PloneMeeting.config import CONFIGURABLE_FIELD_NAMES
 from Products.PloneMeeting.config import CONSIDERED_NOT_GIVEN_ADVICE_VALUE
 from Products.PloneMeeting.config import DEFAULT_COPIED_FIELDS
 from Products.PloneMeeting.config import DUPLICATE_AND_KEEP_LINK_EVENT_ACTION
@@ -1572,6 +1573,23 @@ schema = Schema((
         vocabulary='listItemInitiators',
     ),
     TextField(
+        name='groupsInChargeNotes',
+        allowable_content_types=('text/html',),
+        widget=RichWidget(
+            condition="python: here.adapted().show_field('groupsInChargeNotes')",
+            description="GroupsInChargeNotes",
+            description_msgid="groups_in_charge_notes_descr",
+            label_msgid="PloneMeeting_label_groupsInChargeNotes",
+            label='Groupsinchargenotes',
+            i18n_domain='PloneMeeting',
+        ),
+        default_content_type="text/html",
+        default_output_type="text/x-html-safe",
+        searchable=True,
+        optional=True,
+        write_permission=View,
+    ),
+    TextField(
         name='inAndOutMoves',
         allowable_content_types=('text/html',),
         widget=RichWidget(
@@ -2658,8 +2676,7 @@ class MeetingItem(OrderedBaseFolder, BrowserDefaultMixin):
     def show_field(self, field_name):
         '''See doc in interfaces.py.'''
         item = self.getSelf()
-        if not item.isDefinedInTool() and \
-            item.attribute_is_used(field_name):
+        if item.attribute_is_used(field_name):
             # evaluate TAL expression
             tool = api.portal.get_tool('portal_plonemeeting')
             cfg = tool.getMeetingConfig(item)
@@ -5000,7 +5017,7 @@ class MeetingItem(OrderedBaseFolder, BrowserDefaultMixin):
 
     def _bypass_write_perm_check_for(self, fieldName):
         """See docstring in interfaces.py"""
-        if fieldName in ['neededFollowUp', 'providedFollowUp']:
+        if fieldName in CONFIGURABLE_FIELD_NAMES:
             item = self.getSelf()
             tool = api.portal.get_tool('portal_plonemeeting')
             cfg = tool.getMeetingConfig(item)
@@ -5035,10 +5052,10 @@ class MeetingItem(OrderedBaseFolder, BrowserDefaultMixin):
             self.adapted()._bypass_write_perm_check_for(fieldName)
         # write_permission is "View" for custom management
         # if bypassWritePermissionCheck is False, make sure write_permission
-        # is no more "View", set it to "Modify portal content"
+        # is no more "View", set it to "Manage portal"
         write_perm = field.write_permission
         if not bypassWritePermissionCheck and write_perm == "View":
-            write_perm = ModifyPortalContent
+            write_perm = ManagePortal
         res = checkMayQuickEdit(
             self,
             bypassWritePermissionCheck=bypassWritePermissionCheck,