+++ src/Products/PloneMeeting/browser/overrides.py
@@ -376,7 +376,7 @@ class PloneMeetingOverviewControlPanel(OverviewControlPanel):
         pm_version = api.env.get_distribution('Products.PloneMeeting')._version
         ps = api.portal.get_tool('portal_setup')
         pm_ps_version = ps.getVersionForProfile('Products.PloneMeeting:default')
-        pm_ps_last_version = ps.getLastVersionForProfile('Products.PloneMeeting:default')[0]
+        pm_ps_last_version = u'.'.join(ps.getLastVersionForProfile('Products.PloneMeeting:default'))
         if pm_ps_last_version != pm_ps_version:
             pm_ps_version = u'⚠⚠⚠ %s/%s ⚠⚠⚠ Please launch upgrade steps!!!' % (pm_ps_last_version, pm_ps_version)
         versions.insert(0, 'PloneMeeting %s (%s)' % (pm_version, pm_ps_version))